requirements.txt
Flask
Flask-Session
psycopg2-binary
SQLAlchemy
Werkzeug
python-dotenv
